Move a few pieces of code around
