+++ src/sentry/workflow_engine/processors/detector.py
@@ -10,6 +10,7 @@ from django.conf import settings
 from django.db.models import Q
 from sentry_redis_tools.retrying_cluster import RetryingRedisCluster
 
+from sentry.eventstore.models import GroupEvent
 from sentry.issues.issue_occurrence import IssueOccurrence
 from sentry.issues.producer import PayloadType, produce_occurrence_to_kafka
 from sentry.issues.status_change_message import StatusChangeMessage
@@ -19,12 +20,31 @@ from sentry.utils import metrics, redis
 from sentry.utils.iterators import chunked
 from sentry.workflow_engine.models import DataConditionGroup, DataPacket, Detector, DetectorState
 from sentry.workflow_engine.processors.data_condition_group import evaluate_condition_group
-from sentry.workflow_engine.types import DetectorGroupKey, DetectorPriorityLevel
+from sentry.workflow_engine.types import DetectorGroupKey, DetectorPriorityLevel, DetectorType
 
 logger = logging.getLogger(__name__)
 REDIS_TTL = int(timedelta(days=7).total_seconds())
 
 
+# TODO Think about moving this to workflow/redis.py
+def get_redis_client() -> RetryingRedisCluster:
+    cluster_key = settings.SENTRY_WORKFLOW_ENGINE_REDIS_CLUSTER
+    return redis.redis_clusters.get(cluster_key)  # type: ignore[return-value]
+
+
+# TODO - cache these by evt.group_id? :thinking:
+def get_detector_by_event(evt: GroupEvent) -> Detector:
+    issue_occurrence = evt.occurrence
+
+    if issue_occurrence is None:
+        detector = Detector.objects.get(project_id=evt.project_id, type=DetectorType.ERROR)
+    else:
+        detector = Detector.objects.get(id=issue_occurrence.evidence_data.get("detector_id", None))
+
+    return detector
+
+
+# TODO Think about another locatin for these data classes.
 @dataclasses.dataclass(frozen=True)
 class DetectorEvaluationResult:
     group_key: DetectorGroupKey
@@ -36,57 +56,6 @@ class DetectorEvaluationResult:
     result: IssueOccurrence | StatusChangeMessage | None = None
     # Event data to supplement the `IssueOccurrence`, if passed.
     event_data: dict[str, Any] | None = None
-
-
-# TODO - Add metrics / logging here
-def process_detectors(
-    data_packet: DataPacket, detectors: list[Detector]
-) -> list[tuple[Detector, dict[DetectorGroupKey, DetectorEvaluationResult]]]:
-    results = []
-
-    for detector in detectors:
-        handler = detector.detector_handler
-
-        if not handler:
-            continue
-
-        # TODO add metric here for detector processing
-        detector_results = handler.evaluate(data_packet)
-
-        for result in detector_results.values():
-            if result.result is not None:
-                create_issue_occurrence_from_result(result)
-
-        if detector_results:
-            # TODO - Add metrics / logging here for successful result
-            results.append((detector, detector_results))
-
-        # Now that we've processed all results for this detector, commit any state changes
-        handler.commit_state_updates()
-
-    return results
-
-
-def create_issue_occurrence_from_result(result: DetectorEvaluationResult):
-    occurrence, status_change = None, None
-    if isinstance(result.result, IssueOccurrence):
-        occurrence = result.result
-        payload_type = PayloadType.OCCURRENCE
-    else:
-        status_change = result.result
-        payload_type = PayloadType.STATUS_CHANGE
-
-    produce_occurrence_to_kafka(
-        payload_type=payload_type,
-        occurrence=occurrence,
-        status_change=status_change,
-        event_data=result.event_data,
-    )
-
-
-def get_redis_client() -> RetryingRedisCluster:
-    cluster_key = settings.SENTRY_WORKFLOW_ENGINE_REDIS_CLUSTER
-    return redis.redis_clusters.get(cluster_key)  # type: ignore[return-value]
 
 
 @dataclasses.dataclass(frozen=True)
@@ -416,3 +385,49 @@ class StatefulDetectorHandler(DetectorHandler[T], abc.ABC):
         if updated_detector_states:
             DetectorState.objects.bulk_update(updated_detector_states, ["active", "state"])
         self.state_updates.clear()
+
+
+def create_issue_occurrence_from_result(result: DetectorEvaluationResult):
+    occurrence, status_change = None, None
+    if isinstance(result.result, IssueOccurrence):
+        occurrence = result.result
+        payload_type = PayloadType.OCCURRENCE
+    else:
+        status_change = result.result
+        payload_type = PayloadType.STATUS_CHANGE
+
+    produce_occurrence_to_kafka(
+        payload_type=payload_type,
+        occurrence=occurrence,
+        status_change=status_change,
+        event_data=result.event_data,
+    )
+
+
+# TODO - Add metrics / logging here
+def process_detectors(
+    data_packet: DataPacket, detectors: list[Detector]
+) -> list[tuple[Detector, dict[DetectorGroupKey, DetectorEvaluationResult]]]:
+    results = []
+
+    for detector in detectors:
+        handler = detector.detector_handler
+
+        if not handler:
+            continue
+
+        # TODO add metric here for detector processing
+        detector_results = handler.evaluate(data_packet)
+
+        for result in detector_results.values():
+            if result.result is not None:
+                create_issue_occurrence_from_result(result)
+
+        if detector_results:
+            # TODO - Add metrics / logging here for successful result
+            results.append((detector, detector_results))
+
+        # Now that we've processed all results for this detector, commit any state changes
+        handler.commit_state_updates()
+
+    return results
